renamed package internals back to mutate
